@@ -8,10 +8,14 @@
     "@testing-library/user-event": "^13.5.0",
     "animate.css": "^4.1.1",
     "emailjs-com": "^3.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.3.1",
+    "react-just-parallax": "^3.1.16",
+    "react-router-dom": "^6.22.3",
     "react-scripts": "5.0.0",
+    "react-tooltip": "^5.26.3",
+    "scroll-lock": "^2.1.5",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -41,6 +45,9 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^3.2.3"
+    "autoprefixer": "^10.4.18",
+    "gh-pages": "^3.2.3",
+    "postcss": "^8.4.37",
+    "tailwindcss": "^3.4.1"
   }
 }
