@@ -1,5 +1,13 @@
-import ReactDOM from 'react-dom';
-import App from './App';
-import './index.css';
+import React from "react";
+import ReactDOM from "react-dom/client";
+import { App } from "./App.jsx";
+import { BrowserRouter as Router } from "react-router-dom";
+import "./index.css";
 
-ReactDOM.render(<App/>, document.querySelector("#root"));
+ReactDOM.createRoot(document.getElementById("root")).render(
+  <React.StrictMode>
+    <Router>
+      <App />
+    </Router>
+  </React.StrictMode>
+);
